Created separate managePages file to store tests relating to manage pages
