+++ app/tests/tests.admin.testcafe.ts
@@ -2,6 +2,7 @@ import { landingNavBar } from './navbar.landing.component';
 import { adminNavBar } from './navbar.admin.component';
 import { signinPage } from './signin.page';
 import { landingPage } from './landing.page';
+import { managePages } from './manage.pages';
 import {
   adminHomePage,
   landingInterestExplorerPage,
@@ -71,11 +72,11 @@ test('Test admin management pages', async (testController) => {
   await signinPage.signin(testController, credentials.admin);
   // Unrolling the loop makes this easier to debug.
   await adminNavBar.gotoMenuPageAndVerify(testController, 'manage', 'students');
-  await adminNavBar.clickFilteredStudentsTabAndVerify(testController);
-  await adminNavBar.clickFilteredAlumniTabAndVerify(testController);
-  await adminNavBar.clickAddNewTabAndVerify(testController);
-  await adminNavBar.clickOtherTabAndVerify(testController);
-  await adminNavBar.clickMatriculateTabAndVerify(testController);
+  await managePages.clickFilteredStudentsTabAndVerify(testController);
+  await managePages.clickFilteredAlumniTabAndVerify(testController);
+  await managePages.clickAddNewTabAndVerify(testController);
+  await managePages.clickOtherTabAndVerify(testController);
+  await managePages.clickMatriculateTabAndVerify(testController);
   await adminNavBar.gotoMenuPageAndVerify(testController, 'manage', 'verification');
   await adminNavBar.gotoMenuPageAndVerify(testController, 'manage', 'reviews');
   await adminNavBar.gotoMenuPageAndVerify(testController, 'manage', 'database');
